refactor(api/v2): extract auth domain into its own package
Phase 4 of the internal/api/v2 split: move the auth domain handlers (login,
OAuth callback, logout, auth status) out of the monolithic package api into a
new internal/api/v2/auth subpackage (package authapi), following the established
Phase 4 pattern (Handler embeds *apicore.Core by pointer; the facade Controller
holds a handler field, constructs it, and calls RegisterRoutes in the
deterministic initRoutes order). This is the FINAL leaf domain of Phase 4.

## Scope: handlers moved, middleware stays

Only the auth DOMAIN HANDLERS registered by the old initAuthRoutes move here:
Login (POST /auth/login), OAuthCallback (GET /auth/callback), Logout
(POST /auth/logout), GetAuthStatus (GET /auth/status), plus their domain-local
helpers/types/constants. The auth MIDDLEWARE (GetAuthMiddleware, PrivateModeAuth,
the trusted-proxy IP extractor, the AuthMiddleware field) already moved to
apicore.Core in Phase 1 and stays there; the handlers reuse AuthMiddleware via
promotion from the embedded Core.

## Pattern

- authapi.Handler embeds *apicore.Core by pointer so the shared helpers
  (HandleError, HandleErrorWithKey, the LogSecurity*IfEnabled / LogDebugIfEnabled
  family, and the AuthMiddleware field) promote onto it.
- New(core, authService) constructs the handler; RegisterRoutes(g *echo.Group)
  wires the exact 4 routes from the old initAuthRoutes in the same order with the
  same per-route middleware: the login rate limiter on POST /login (5 attempts /
  15 min per IP, byte-identical config) and AuthMiddleware on the protected
  logout/status group (authGroup.Group("", c.AuthMiddleware)).
- Handler methods moved verbatim, only the receiver retyped from *Controller to
  *Handler (receiver name kept c). Bodies read c.authService (the Handler's own
  field) exactly as before.

## Dependency wiring (the auth-specific part)

authService (auth.Service) is INJECTED from the facade. Unlike the other domain
deps, which are set in the Controller literal before construction, authService is
supplied via the WithAuthService functional option, which is applied in the
`for _, opt := range opts { opt(c) }` loop AFTER the other handlers are built. So
the facade constructs c.authHandler = authapi.New(c.Core, c.authService) AFTER
that options loop, guaranteeing the handler captures the injected service (or nil
if WithAuthService was not passed; the handlers nil-guard authService exactly as
the monolith did). authService never changes after the options loop, so capturing
the post-option value is behaviorally identical to the monolith's per-request
`c.authService` read. The facade keeps the authService field and the
WithAuthService option as the injection source.

## Cross-domain coupling

- The 5 auth route-path constants (AuthGroupPath..AuthStatusPath) are exported in
  authapi and consumed by the facade's isPrivateModeExempt PrivateMode allow-list
  (audio_hls.go) and its tests, so the exempt list composes its auth paths from
  the same constants the routes register with and cannot drift (the same approach
  as RangeFilterSpecies in rangeapi and ParseCertificateInfo in tls). The const
  expression `const authBase = apiV2Prefix + authapi.AuthGroupPath` stays a valid
  compile-time constant.
- fuzz_test.go (package api) is a mixed security file: some test functions call
  both auth validators and non-auth validators (NormalizeClipPathStrict,
  parseConfidenceFilter) in the same loop, so they cannot all move. The 6
  dedicated pure-auth Fuzz functions + TestAdvancedRedirectAttacks moved verbatim
  to authapi/auth_fuzz_test.go (auth now owns its security fuzzing in its own
  -race binary). Three validators (IsValidBasePath, ValidateAndSanitizeRedirect,
  ContainsCRLFCharacters) are exported so the remaining facade mixed tests
  exercise them across the package boundary; their bodies are unchanged (only the
  identifier case differs), so coverage is preserved with zero loss.

## Facade wiring (order preserved exactly)

- Controller gains an unexported authHandler *authapi.Handler field.
- c.authHandler = authapi.New(c.Core, c.authService) is constructed after the
  options loop.
- initRoutes replaces the auth routeInitializers entry in place with
  {"auth routes", func() { c.authHandler.RegisterRoutes(c.Group) }} at the same
  index, so registration order (and the route-enumeration golden gate) is
  byte-identical.

auth owns no background goroutine or singleton, so no Shutdown hook is added; the
facade Shutdown is untouched. apiv2.Controller/New/NewWithOptions signatures are
unchanged.

## Tests

Moved into the auth package, rebuilt against apitest:
- auth_compatibility_test.go (self-contained V1/V2 auth-logic simulation).
- auth_integration_test.go: setup rewritten to build the Handler via
  apitest.NewCore(...) + New(core, authService) with a real SecurityAdapter,
  invoking the handlers directly (same assertions: full login+callback flow,
  filtered-query redirect survival, proxy-prefixed base path, invalid/missing/
  wrong credentials, empty-ClientID V1 compatibility, callback errors, open-
  redirect prevention, Service-interface conformance).
- auth_fuzz_test.go (the moved fuzz funcs + advanced redirect attacks).
- auth_routes_test.go (route-enumeration gate for the 4 auth routes; auth now has
  its own -race binary).

## Verification

- go build ./... clean.
- go vet ./internal/api/v2/... ./internal/analysis/... clean (no copylocks).
- go test -race ./internal/api/v2/... green (incl. the route-enumeration golden
  gate and the PrivateMode-exempt tests); internal/api/v2/auth has its own -race
  binary; internal/analysis -race green (its only failure was the unrelated
  frontend dist embed, which is environmental).
- golangci-lint: 0 issues on internal/api/v2/... and internal/api/v2/auth/.
- No public-behavior change: auth endpoints, auth/CSRF/session/rate-limit/redirect
  handling, and responses are identical; README endpoint catalog accurate.

## Security preservation

Every auth check is preserved verbatim: the login rate limiter config and its
error/deny handlers, the AuthMiddleware on the protected logout/status routes,
the timing-attack randomDelay on the missing-credential and auth-failure paths,
the query-aware open-redirect validation (path/query split, security.IsValidRedirect,
EscapedPath emission), the session-fixation EstablishSession, and fetching the
provider logout URL before Logout() destroys the session.
